Datasets/evalplus/Java/src/test/java/com/example/HumanEval_68Test.java
package com.example;
import static org.junit.Assert.*;
import org.junit.Test;

public class HumanEval_68Test {

    @Test
    public void testPluck0() {
        int[] arr = {4, 2, 3};
        int[] result = HumanEval_68.pluck(arr);
        assertArrayEquals(new int[]{2, 1}, result);
    }

    @Test
    public void testPluck1() {
        int[] arr = {1, 2, 3};
        int[] result = HumanEval_68.pluck(arr);
        assertArrayEquals(new int[]{2, 1}, result);
    }

    @Test
    public void testPluck2() {
        int[] arr = new int[0];
        int[] result = HumanEval_68.pluck(arr);
        assertArrayEquals(new int[]{}, result);
    }

    @Test
    public void testPluck3() {
        int[] arr = {5, 0, 3, 0, 4, 2};
        int[] result = HumanEval_68.pluck(arr);
        assertArrayEquals(new int[]{0, 1}, result);
    }

    @Test
    public void testPluck4() {
        int[] arr = {1, 2, 3, 0, 5, 3};
        int[] result = HumanEval_68.pluck(arr);
        assertArrayEquals(new int[]{0, 3}, result);
    }

    // Continue with other test cases...
}
// End of Test
